agent_pipeline/requirements.txt
crewai==0.51.1
langchain-openai
jinja2==3.1.3
pydantic
python-dotenv==1.0.1
